Initial working checkin
This includes a working LCM and one which we can not crack yet.
The working one, we get the 'FULL' seed values returned.  In the broken one (VC's rand() function), we only get the the high 16 bits (with high bit cut off).  This does not work in the cracker yet, but it will have to be made TO work, so that we can later work on finding the MySpace RNG string generation code.